django_project/requirements.txt
django==1.11.3
django-debug-toolbar==1.8